feat: refacto, changing from int to long
