+++ include/libft.h
@@ -228,6 +228,15 @@ char	*ft_strnstr(char *str, char *search, size_t size);
 int		ft_atoi(const char *str);
 
 /**
+ @brief Transform a string into a long with support for different bases.
+ @param str A string.
+ @return A long integer corresponding to the digits of the string.
+ Supports decimal (default), hexadecimal (0x prefix), binary (0b prefix),
+ and octal (0 prefix) bases.
+*/
+long	ft_atoi_base(char *str);
+
+/**
  @brief Duplicate a given string.
  @param str A string.
  @return A pointer to a new string allocated in the memory that is identical
@@ -285,6 +294,14 @@ char	**ft_split(char const *str, char c);
  the integer given.
 */
 char	*ft_itoa(int nb);
+
+/**
+ @brief Transform a long integer into a string.
+ @param nb A long integer.
+ @return A new string allocated containing the digits that correspond to
+ the long integer given.
+*/
+char	*ft_ltoa(long nb);
 
 /**
  @brief Create a new string where each character from the old string is passed
