+++ lib/Makefile
@@ -4,8 +4,8 @@ SHELL = bash
 
 FILES =	is/ft_isdigit is/ft_isalnum is/ft_isalpha is/ft_isascii is/ft_isprint \
 		mem/ft_memset mem/ft_bzero mem/ft_memcpy mem/ft_memmove mem/ft_memchr mem/ft_memcmp mem/ft_calloc \
-		str/ft_strlen str/ft_strlcpy str/ft_strlcat str/ft_toupper str/ft_tolower str/ft_strchr str/ft_strrchr str/ft_strncmp str/ft_strnstr str/ft_atoi \
-		str/ft_strdup str/ft_strcmp str/ft_substr str/ft_strjoin str/ft_strjoin_free str/ft_strtrim str/ft_split str/ft_itoa str/ft_strmapi str/ft_striteri \
+		str/ft_strlen str/ft_strlcpy str/ft_strlcat str/ft_toupper str/ft_tolower str/ft_strchr str/ft_strrchr str/ft_strncmp str/ft_strnstr str/ft_atoi str/ft_atoi_base \
+		str/ft_strdup str/ft_strcmp str/ft_substr str/ft_strjoin str/ft_strjoin_free str/ft_strtrim str/ft_split str/ft_itoa str/ft_ltoa str/ft_strmapi str/ft_striteri \
 		put/ft_putchar_fd put/ft_putstr_fd put/ft_putendl_fd put/ft_putnbr_fd put/ft_puthexa_fd put/ft_putunbr_fd \
 		lst/ft_lstnew lst/ft_lstadd_back lst/ft_lstadd_front lst/ft_lstclear lst/ft_lstdelone lst/ft_lstiter lst/ft_lstmap lst/ft_lstsize lst/ft_lstlast \
 		printf/ft_printf printf/ft_fprintf get_next_line/get_next_line_utils get_next_line/get_next_line
